Change app name & remove a link
Finally gave the app a name and it's reflected in the repo and the
app. Also took away the link to gcor.io because the site is kind
of messy and doesn't contribute anything to the application. A link
may be put back there later when the site has something else of use
to visitors of the application.