Re-structure repo & add image files
The repo was originally put structured just to provide code that
supplies logic for the Clarity app. But for anyone that wants to
use the application locally, there were files missing and the
repo was structured in such a way where Flask would not know
how to access several files.

Those issues should be resolved with this commit.