Refactor application code and improve template readability
No changes to user experience in this commit.

The program was not buggy but it was messy in its structure. I feel
it was not reader-friendly to someone not familiar with the
application's code. This commit changes it to take a class-based
approach, which has the immediate effect of making each individual
python file easier to digest as well as making the html templates
easier to understand.

Files added:
- player.py -- This file defines the Player class
- constants.py -- This file holds the constants needed by the app

Files modified:
- app.py -- Readability improved, makes use of the Player class now
instead of handling all the application logic on its own
- search.html -- Readability improved by using Player class attributes
- summoner.html -- Readability improved by using Player class
attributes